Refactor: Se movieron los archivos encargados de la conexion a la bd y de mapear las clases en tablas a la carpeta src
