+++ app.js
@@ -7,7 +7,7 @@ app.use(morgan("dev"));
 
 // Carga la configuración desde config.json
 const env = process.env.NODE_ENV || "development";
-const config = require("./config/config.json")[env];
+const config = require("./src/config/config.json")[env];
 
 // Configura la conexión a la base de datos MySQL
 const sequelize = new Sequelize(
@@ -22,11 +22,11 @@ const sequelize = new Sequelize(
 
 // Importa los modelos Sequelize
 
-const PersonaModel = require("./models/Persona");
-const LocalidadModel = require("./models/Localidad");
-const ProvinciaModel = require("./models/Provincia");
-const PublicacionModel = require("./models/Publicacion");
-const ServicioModel = require("./models/Servicio");
+const PersonaModel = require("./src/models/Persona");
+const LocalidadModel = require("./src/models/Localidad");
+const ProvinciaModel = require("./src/models/Provincia");
+const PublicacionModel = require("./src/models/Publicacion");
+const ServicioModel = require("./src/models/Servicio");
 
 // Define los modelos en Sequelize
 const Persona = PersonaModel(sequelize, Sequelize.DataTypes);
@@ -36,7 +36,7 @@ const Publicacion = PublicacionModel(sequelize, Sequelize.DataTypes);
 const Servicio = ServicioModel(sequelize, Sequelize.DataTypes);
 
 // Configura las asociaciones
-require("./models/associations")(sequelize);
+require("./src/models/associations")(sequelize);
 
 // Sincroniza los modelos con la base de datos
 sequelize
